store onboaring

+++ db/migrate/20221110074143_create_store_mdrs.rb
@@ -4,7 +4,7 @@ class CreateStoreMdrs < ActiveRecord::Migration[5.2]
   def change
     create_table :store_mdrs do |t|
       t.string  :payment_mode
-      t.integer :rate_amount
+      t.float   :rate_amount
       t.string  :rate_type
       t.integer :upto_amount
       t.integer :from_amount
