+++ db/migrate/20221110074345_create_bank_mdrs.rb
@@ -4,7 +4,7 @@ class CreateBankMdrs < ActiveRecord::Migration[5.2]
   def change
     create_table :bank_mdrs do |t|
       t.string  :payment_mode
-      t.integer :rate_amount
+      t.float   :rate_amount
       t.string  :rate_type
       t.integer :upto_amount
       t.integer :from_amount
